+++ R/census.R
@@ -82,6 +82,83 @@ node_triad_census <- function(object){
 }
 
 #' @rdname census
+#' @importFrom oaqc oaqc
+#' @importFrom tidygraph %E>%
+#' @examples 
+#' (quad_cen <- node_quad_census(southern_women))
+#' @export
+node_quad_census <- function(object){
+  graph <- object %>% as_tidygraph() %E>% 
+    as.data.frame()
+  out <- oaqc::oaqc(graph)[[1]]
+  out <- out[-1,]
+  rownames(out) <- node_names(object)
+  colnames(out) <- c("E4", # co-K4
+                     "I41","I40", # co-diamond
+                     "H4", # co-C4
+                     "L42","L41","L40", # co-paw
+                     "D42","D40", # co-claw
+                     "U42","U41", # P4
+                     "Y43","Y41", # claw
+                     "P43","P42","P41", # paw
+                     "04", # C4
+                     "Z42","Z43", # diamond
+                     "X4") # K4
+  if(is_twomode(object)) out <- out[,-c(8,9,14,15,16,18,19,20)]
+  out
+}
+
+#' @rdname graph_census
+#' @param object2 A second, two-mode migraph-consistent object
+#' @references 
+#' Hollway, James, Alessandro Lomi, Francesca Pallotti, and Christoph Stadtfeld. 
+#' “\href{https://doi.org/10.1017/nws.2017.8}{Multilevel Social Spaces: The Network Dynamics of Organizational Fields}.” 
+#' _Network Science_ 5, no. 2 (June 2017): 187–212.
+#' @source Alejandro Espinosa 'netmem'
+#' @examples 
+#' marvel_friends <- to_unsigned(ison_marvel_relationships, "positive")
+#' (mixed_cen <- graph_mixed_census(marvel_friends, ison_marvel_teams))
+#' @export
+graph_mixed_census <- function (object, object2) {
+  if(is_twomode(object)) stop("First object should be a one-mode network")
+  if(!is_twomode(object2)) stop("Second object should be a two-mode network")
+  if(graph_dimensions(object)[1]!=graph_dimensions(object2)[1]) stop("Non-conformable arrays")
+  
+  m1 <- as_matrix(object)
+  m2 <- as_matrix(object2)
+  cp <- function(m) (-m + 1)
+  
+  onemode.reciprocal <- m1 * t(m1)
+  onemode.forward <- m1 * cp(t(m1))
+  onemode.backward <- cp(m1) * t(m1)
+  onemode.null <- cp(m1) * cp(t(m1))
+  diag(onemode.forward) <- 0
+  diag(onemode.backward) <- 0
+  diag(onemode.null) <- 0
+  
+  bipartite.twopath <- m2 %*% t(m2)
+  bipartite.null <- cp(m2) %*% cp(t(m2))
+  bipartite.onestep1 <- m2 %*% cp(t(m2))
+  bipartite.onestep2 <- cp(m2) %*% t(m2)
+  diag(bipartite.twopath) <- 0
+  diag(bipartite.null) <- 0
+  diag(bipartite.onestep1) <- 0
+  diag(bipartite.onestep2) <- 0
+  
+  res <- c("22" = sum(onemode.reciprocal * bipartite.twopath) / 2,
+           "21" = sum(onemode.forward * bipartite.twopath) / 2 + sum(onemode.backward * bipartite.twopath) / 2,
+           "20" = sum(onemode.null * bipartite.twopath) / 2,
+           "12" = sum(onemode.reciprocal * bipartite.onestep1) / 2 + sum(onemode.reciprocal * bipartite.onestep2) / 2,
+           "11D" = sum(onemode.forward * bipartite.onestep1) / 2 + sum(onemode.backward * bipartite.onestep2) / 2,
+           "11U" = sum(onemode.forward * bipartite.onestep2) / 2 + sum(onemode.backward * bipartite.onestep1) / 2,
+           "10" = sum(onemode.null * bipartite.onestep2) / 2 + sum(onemode.null * bipartite.onestep1) / 2,
+           "02" = sum(onemode.reciprocal * bipartite.null) / 2,
+           "01" = sum(onemode.forward * bipartite.null) / 2 + sum(onemode.backward * bipartite.null) / 2,
+           "00" = sum(onemode.null * bipartite.null) / 2)  
+  return(res)
+}
+
+#' @rdname census
 #' @examples 
 #' group_tie_census(task_eg, cutree(cluster_structural_equivalence(task_eg), 4))
 #' @export
@@ -124,17 +201,6 @@ group_triad_census <- function(object, clusters, decimals = 2) {
     rownames(cluster_triad_mat) <- clusters
   }
   cluster_triad_mat 
-}
-
-#' @title Get triad census summary by group
-#' `r lifecycle::badge("deprecated")`
-#' @keywords internal
-#' @description Deprecated on 2021-10-26.
-#' @return `group_triad_census()`
-#' @export
-cluster_triad_census <- function(object, clusters) {
-  .Deprecated("group_triad_census")
-  group_triad_census(object, clusters)
 }
 
 #' Censuses for the whole graph