+++ R/ggidentify.R
@@ -16,6 +16,7 @@ NULL
 #' ggidentify(brandes, node_eigenvector)
 #' @export
 ggidentify <- function(object, node_measure, identify_function = max) {
+  object <- as_tidygraph(object)
   measure <- node_measure(object)
   colord <- ifelse(measure == identify_function(measure),
                "max", "other")
@@ -38,10 +39,14 @@ ggidentify <- function(object, node_measure, identify_function = max) {
 #' @export
 ggdistrib <- function(object, node_measure){
   distrib <- node_measure(object)
-  ggplot2::qplot(distrib, geom="histogram", 
-                 binwidth=ifelse(max(distrib)>1, 1, 
-                                 ifelse(max(distrib)>.1, .1, .01))) + 
+  ggplot2::ggplot(as.data.frame(distrib)) +
+    ggplot2::geom_histogram(ggplot2::aes(distrib),
+                            binwidth = ifelse(max(distrib) > 1, 1,
+                                              ifelse(max(distrib) > .1,
+                                                     .1,
+                                                     .01))) +
     ggplot2::theme_classic() +
+    ggplot2::theme(panel.grid.major = ggplot2::element_line(colour = "grey90")) +
     ggplot2::xlab("Score") +
     ggplot2::ylab("Frequency")
 }