+++ tests/testthat/test-ggidentify.R
@@ -32,3 +32,16 @@ test_that("Plot identify function works", {
   expect_named(testDegree[1], "data")
   expect_true(is.call(testEigen[["layers"]][[2]][["mapping"]][["colour"]]))
 })
+
+testDegree     <- ggdistrib(brandes, node_degree)
+testBetweeness <- ggdistrib(brandes, node_betweenness)
+testCloseness  <- ggdistrib(brandes, node_closeness)
+testEigen      <- ggdistrib(brandes, node_eigenvector)
+
+test_that("Plot identify function works", {
+  expect_true(is.list(testEigen))
+  expect_length(testDegree, 9)
+  expect_named(testDegree[1], "data")
+  expect_equal(testEigen[["labels"]][["y"]], "Frequency")
+  expect_equal(testBetweeness[["labels"]][["x"]], "Score")
+})